Fix issue with some dir listings being allocated in multiple attributes.

+++ NTFSLib/NTFS.cs
@@ -83,7 +83,7 @@ namespace NTFSLib
             Debug.Assert(fileMftData.DataFragments.Length >= 1);
 
             // Get number of FileRecords 
-            FileRecordCount = (uint)(fileMftData.DataFragments.Sum(s => ((int)s.Clusters * (int)BytesPrCluster)) / BytesPrFileRecord);
+            FileRecordCount = (uint)(fileMftData.DataFragments.Sum(s => (float)s.Clusters) * (BytesPrCluster / BytesPrFileRecord));
             FileRecords = new WeakReference[FileRecordCount];
 
             FileRecords[0] = new WeakReference(FileMFT);
